Don't announce a logical size if it's the same it already had
I only found this when I started hitting a bug that was calling this
method repeatedly. I figured it still makes sense as it's a lot of
traffic that in this case won't ultimately result in meaningful
information.

This prevents from overwhelming clients that use libkscreen from
starting to do nothing repeatedly.