Add high-resolution scrolling support
In the wl_pointer version 8 interface, the axis_value120 replaces the
axis_discrete event.

Conceptually, with the axis_value120 event, a single wheel click is
broken up in multiple smaller clicks. 120 corresponds to a single
discrete click.

In order to keep clients that don't support version 8 yet, kwin
accumulates delta v120 values and when the accumulator reaches a
multiple of 120, it will send the axis_discrete event.

https://gitlab.freedesktop.org/wayland/wayland/-/merge_requests/72