Allow special clients to receive all keyboard input
Certain technologies that we mean to support, namely X11, are designed
with the premise that clients can eavesdrop the users' input. If we want
to fully support them, we need to offer such information so apps keep
working.

This specifically will allow us to support, among others, Push-to-Talk
implementations that like electron's iohook that just expect to filter
all user input themselves instead of actively requesting what they need.