WIP: Introduce client level scale overrides
This allows for compositor managed different co-ordinate space between
kwin's logical co-ordinate space and a client's logical co-ordinate
space which opens a range of flexibilty.

When combined with a modified
https://invent.kde.org/plasma/kwin/-/merge_requests/1959 this allows us
to set a DPI in xrdb and mark all xwayland windows as being 2x (or
other) and avoid upscales for xwayland clients in a way that doesn't
impact other wayland clients or require third-party changes. Any use of
fracitonal values is in layers we control.

kwayland-server is the right place for this abstraction as we need
Outputs to differ on a per resource basis. Something we can't control
from within kwin.

If we covered remaining protocols we can offer user-control on all
remaining clients which could open up some interesting possibilities.